rfp/requirements_server.txt
numpy==1.23.4
scipy==1.9.3
pandas~=1.5.3
protobuf==3.20.3
python-dotenv~=1.0.1
streamlit==1.26.0
google-auth==1.35.0
openai>=0.28.1
langid==1.1.6
typing-inspect==0.8.0
typing_extensions>=4.5.0
pydantic==1.10.11
langchain==0.0.200
streamlit-toggle-switch
# hotel pdf parser
# pdf2image~=1.16.3
stqdm==0.0.5
requests~=2.32.3
pinecone~=5.3.1
pinecone-text
flask~=3.0.2
# Pillow~=9.3.0
sentence-transformers
nltk
